Fix expectation of array literals with trailing empty elements.
If the last empty elements are not specified (as undefined), then the
expected syntax tree will be falsified because the array was not
properly encoded. The solution is to give explicit suffix for those
empty elements.

This fixes some false positives in the compatibility tests, bringing down the
total failures to only 78 (90% pass rate).

http://code.google.com/p/esprima/issues/detail?id=5