Avoid recursion in most binary expressions.
An expression which consists of a series of binary expressions with the
same operator precedence can be processed linearly, without falling back
to recursive parsing.

Example expression:

    1 + 2 + 3

should be handled as if it looks like:

    (1 + 2) + 3

and not like:

    1 + (2 + 3)

New unit tests have been added to cover these cases.

This solves the remaining failures in the compatibility tests. It's now
100% pass rate!

http://code.google.com/p/esprima/issues/detail?id=58