Fix automatic semicolon insertion.
Mostly based on patch from Yusuke Suzuki <utatane.tea@gmail.com>.

It's now handled for the following statements:

  continue
  break
  return
  throw

and the following postfix expressions:

  ++
  --

as well the fix for consumeSemicolon() function.

Extensive tests are added to cover this case.
Unfortunately we have to suffer performance loss (9% with Chrome 15 on
the benchmarks suite). This is the price we pay for being correct.

http://code.google.com/p/esprima/issues/detail?id=35