Add support for optional statement-level comment node in the syntax tree.
By specifying 'comment' to true in the parse() function, there will be
an additional Comment node in the syntax tree, for every line comment
and block comment.

Currently it works only in these cases:
  top-level statements, i.e. SourceElements
  statements inside block { ... }
  case statements in a switch

New tests have been added to cover the above cases. In addition, there is
a special code path for comments handling, the performance when not including
the comments is not affected.

The parser demo is extended with a checkbox to specify whether to
include the comments (in the resulting syntax tree) or not.

http://code.google.com/p/esprima/issues/detail?id=71