Add location info to the exception.
The location is expressed both in index (0-based) and lineNumber/column
(1-based). The former is useful to refer to the original source string
while the latter is more for the user info.

The tests fixture is modified to check for the location info as well.

The error message is kept to mention the line number for convenience,
calling toString() on the exception object should show human-friendly
message string already.

The column is obtained by storing the index where each line starts.
This part is from Kris Kowal <kris.kowal@cixar.com>.

http://code.google.com/p/esprima/issues/detail?id=140