Fix to match parenthesis of wrapped member or call expressions.
The character range of MemberExpression in "(a.b)" was "a.b)".
The commit fixes this problem by returning the range of "(a.b)"
in these cases.

Two of already existing unit tests were expecting wrong result
for inputs like "(function () { ... }())". These are updated to
expect the full range of expression.

Few additional test cases are added to test the parsing of
similar expressions and cover all the added and modified lines.

http://code.google.com/p/esprima/issues/detail?id=320