Fix incorrect range of expressions inside parentheses
The character range of expressions in parentheses have included
the parentheses as well.

An example: "(42)"
Incorrect range of literal:   [0,4]
The correct range of literal: [1,3]

Furthermore in case of MemberExpression and CallExpression the
range have included the closing parenthesis but not the opening.
The commit fixes that as well.

An example: "(f())"
Incorrect range of call expression:   [1,5]
The correct range of call expression: [1,4]

13 tests have been updateded to expect the correct range.

http://code.google.com/p/esprima/issues/detail?id=309