Change member expression parsing.
Because of the nature of brackets like [] and (), we can't use run-time
patching to capture the position before and after some member expression
parsing routines. The solution is to create a specialized version when
we need to track the syntax node location.

Coverage regression is expected (temporarily) until variation in the
parsing options when running the test suite gets implemented.

http://code.google.com/p/esprima/issues/detail?id=324
http://code.google.com/p/esprima/issues/detail?id=343

Conflicts:
	esprima.js