Treat yield as a contextual keyword.
This means, in particular, that `yield` is a perfectly valid identifier
outside the bodies of generator functions, and that it is always treated
as an operator inside the bodies of generator functions.

When we encounter an expression like `yield x` outside a generator
function, parsing fails with an "Unexpected identifier" error pointing to
the `x`, as you would expect with `whatever x`.

https://code.google.com/p/esprima/issues/detail?id=467