No null in ArrowFunctionExpression defaults
`FunctionExpression` and `FunctionDeclaration` both use `undefined` as a placeholder in their `defaults` list for parameters without a default value. This follows the SpiderMonkey spec which specifies `defaults: [ Expression ];` `ArrowFunctionExpression`, however, is using `null` as a placeholder in its `defaults` list. This is inconsistent and doesn't match the SpiderMonkey spec (though I do understand that matching the SpiderMonkey spec with regards to `ArrowExpression` is a non-goal)