Expand property method range and loc to include params
Previously, the range and loc for get, set, and shorthand methods
inside object literal properties included only the method body,
meaning a method's parameters were outside the range of their parent
`FunctionExpression` node. This was happening because the
`FunctionExpression` node wasn't being created until
`parsePropertyFunction` was called, by which point the method's
parameters had already been parsed. This PR solves the issue by
creating the `FunctionExpression` node as soon as it determines that
the property is a get, set, or shorthand method, then passing the
node into `parsePropertyFunction`.

Closes #1029
Closes #1040