Run the tests of downstream projects.
This is done by checking out each project, run npm install, drop-in a
replacement for esprima.js, and run npm test. As of now, this is carried
out only as part of Travis CI build. It can be tested manually by:

  rm -rf downstream
  npm run downstream

Due to some unknown reason, .eslintrc can't exist otherwise it is going
to confuse the linting of the downstream project. The workaround is to
use custom config file.

Refs #1023