Add Karma TestRunner
Update Unit Tests to use karma test-runner. The change, is largely
because we'd like to be able to run the unit tests in the browser
and it's nice to have a configurable test runner.

There were a couple changes to the unit test build order:

**generateTestCase** was moved to it's own tool. `tool/generate-test-case`

**Fixtures**: Fixtures JSON and JS files are preprocessed and loaded via <script>
tags.

The JSON fixtures are available via at `__mocks__` and are processed
with `karma-json-processor`. The JS fixtures are available via
`__fixtures` and were processed with `karma-js2js-processor` that I
wrote. It saves the JS in a multiline comment that needs to be unwrapped
after page load.

The enumerateFixtures function was updated to account for
the changes.

**Mocha** The custom reporter/runner was replaced with
mocha/chai/sinon/leche, which has
a couple of benefits. At the moment, we're mostly benefiting from the
pretty reporter. Down the road, I anticipate this setup allowing the
tests to be more expressive.

Next Steps / Questions / Concerns:
+ There are some failing tests that I'd like some help hunting down. I
believe they're unicode related
+ The tests are organized based on the testX functions (testParse,
testModule, testTree, etc). This made sense at the time, but now i'm
wondering if it'd make more sense to organize the test files in terms of
the fixtures (api, comment, ES6) going forward. If so, we can do this in
a next step.
+ The test run time is significantly slower now. Previously, the tests
ran in 1.3 seconds. Now, the tests run in the same amount of time, but
there's a 15 second load time for the script tags. This could be okay,
as karma has a helpful watch feature, but could also be a workflow pain
point.