Normalize regex testing when using the new ES6 regex flags.
On an environment that already supports the new ES6 regex flags,
the regex value won't be null. This causes the test to be mistakenly
marked as a failure since the baseline has the null value.

To workaround this, ensure that regex value is always null when using
the new ES6 regex flags. Other comparisons will still happen to ensure
that the regex is handled properly.

Fixes #1271.
Closes gh-1274