Revamp most of the parsing code
The bulk of the code is encapsulated in a new class, Parser. For each
parsing function, the parameters and the return value are annotated with
the corresponding type information. As such, the AST constructors are
also getting the additional type information.

Some outdated micro-optimizations not necesary anymore (as evidenced
from the benchmark tests) are removed.

Fixes #1410
Closes gh-1413